+++ massiv-test/tests/Data/Massiv/Array/Ops/ConstructSpec.hs
@@ -107,7 +107,7 @@ prop_unfoldrList sz1 f i =
   conjoin $
   L.zipWith
     (===)
-    (A.toList (computeAs P $ unfoldrS_ Seq sz1 (apply f) i))
+    (A.toList (computeAs P $ unfoldrS_ sz1 (apply f) i))
     (L.unfoldr (Just . apply f) i)
 
 specExpand :: Spec