+++ massiv-test/tests/Data/Massiv/Array/Ops/SliceSpec.hs
@@ -107,33 +107,36 @@ prop_SliceIndexDim3RankD (ArrIx arr ix@(i :> j :. k)) =
     val = unsafeIndex arr ix
 
 
-prop_SliceIndexDim2M :: ArrIx M Ix2 Int -> Property
-prop_SliceIndexDim2M (ArrIx arr ix@(i :. j)) =
+prop_SliceIndexDim2M :: ArrIx P Ix2 Int -> Property
+prop_SliceIndexDim2M (ArrIx arr' ix@(i :. j)) =
   val === (arr !> i ! j) .&&.
   val === (arr <! j ! i)
   where
+    arr = toManifest arr'
     val = unsafeIndex arr ix
 
-prop_SliceIndexDim2RankM :: ArrIx M Ix2 Int -> Property
-prop_SliceIndexDim2RankM (ArrIx arr ix@(i :. j)) =
+prop_SliceIndexDim2RankM :: ArrIx P Ix2 Int -> Property
+prop_SliceIndexDim2RankM (ArrIx arr' ix@(i :. j)) =
   val === (arr <!> (2, i) ! j) .&&.
   val === (arr <!> (1, j) ! i)
   where
+    arr = toManifest arr'
     val = unsafeIndex arr ix
 
 
-prop_SliceIndexDim3M :: ArrIx M Ix3 Int -> Property
-prop_SliceIndexDim3M (ArrIx arr ix@(i :> j :. k)) =
+prop_SliceIndexDim3M :: ArrIx P Ix3 Int -> Property
+prop_SliceIndexDim3M (ArrIx arr' ix@(i :> j :. k)) =
   val === (arr <! k <! j ! i) .&&.
   val === (arr !> i !> j ! k) .&&.
   val === (arr <! k !> i ! j) .&&.
   val === (arr !> i <! k ! j)
   where
+    arr = toManifest arr'
     val = unsafeIndex arr ix
 
 
-prop_SliceIndexDim3RankM :: ArrIx M Ix3 Int -> Property
-prop_SliceIndexDim3RankM (ArrIx arr ix@(i :> j :. k)) =
+prop_SliceIndexDim3RankM :: ArrIx P Ix3 Int -> Property
+prop_SliceIndexDim3RankM (ArrIx arr' ix@(i :> j :. k)) =
   val === (arr <!> (3, i) <!> (2, j) ! k) .&&.
   val === (arr <!> (3, i) <!> (1, k) ! j) .&&.
   val === (arr <!> (2, j) <!> (2, i) ! k) .&&.
@@ -141,6 +144,7 @@ prop_SliceIndexDim3RankM (ArrIx arr ix@(i :> j :. k)) =
   val === (arr <!> (1, k) <!> (2, i) ! j) .&&.
   val === (arr <!> (1, k) <!> (1, j) ! i)
   where
+    arr = toManifest arr'
     val = unsafeIndex arr ix
 
 
@@ -169,8 +173,8 @@ prop_SliceIndexDim4RankD (ArrIx arr ix@(i1 :> i2 :> i3 :. i4)) =
     val = evaluate' arr ix
 
 
-prop_SliceIndexDim4RankM :: ArrIx M Ix4 Int -> Property
-prop_SliceIndexDim4RankM (ArrIx arr ix@(i1 :> i2 :> i3 :. i4)) =
+prop_SliceIndexDim4RankM :: ArrIx P Ix4 Int -> Property
+prop_SliceIndexDim4RankM (ArrIx arr' ix@(i1 :> i2 :> i3 :. i4)) =
   val === (arr <!> (4, i1) <!> (3, i2) <!> (2, i3) ! i4) .&&.
   val === (arr <!> (4, i1) <!> (2, i3) <! i4 ! i2) .&&.
   val === (arr <!> (3, i2) <!> (3, i1) ! (i3 :. i4)) .&&.
@@ -178,11 +182,12 @@ prop_SliceIndexDim4RankM (ArrIx arr ix@(i1 :> i2 :> i3 :. i4)) =
   val === (arr <!> (2, i3) <!> (1, i4) !> i1 ! i2) .&&.
   val === (arr <!> (1, i4) !> i1 !> i2 ! i3)
   where
+    arr = toManifest arr'
     val = unsafeIndex arr ix
 
 
-prop_SliceIndexDim4M :: ArrIx M Ix4 Int -> Property
-prop_SliceIndexDim4M (ArrIx arr ix@(i1 :> i2 :> i3 :. i4)) =
+prop_SliceIndexDim4M :: ArrIx P Ix4 Int -> Property
+prop_SliceIndexDim4M (ArrIx arr' ix@(i1 :> i2 :> i3 :. i4)) =
   val === (arr !> i1 !> i2 !> i3 ! i4) .&&.
   val === (arr !> i1 !> i2 <! i4 ! i3) .&&.
   val === (arr !> i1 <! i4 <! i3 ! i2) .&&.
@@ -192,6 +197,7 @@ prop_SliceIndexDim4M (ArrIx arr ix@(i1 :> i2 :> i3 :. i4)) =
   val === (arr <! i4 <! i3 <! i2 ! i1) .&&.
   val === (arr <! i4 <! i3 !> i1 ! i2)
   where
+    arr = toManifest arr'
     val = unsafeIndex arr ix